# Alignment with ksw2 and WFA2-lib