… CMakeLists.txt
… ext/README.md
… ext/ksw2/LICENSE.txt
… ext/ksw2/ksw2.h
… ext/ksw2/ksw2_extz2_sse.c
… ext/ksw2_extz2_sse.c
… ext/ssw/ssw.c
… ext/ssw/ssw.h
… ext/ssw/ssw_cpp.cpp
… ext/ssw/ssw_cpp.h
… src/aln.cpp
… src/aln.hpp
… src/cigar.cpp
… src/cigar.hpp
… src/ksw2wrap.cpp
… src/ksw2wrap.hpp
… src/pc.hpp
… tests/test_aligner.cpp
… tests/test_cigar.cpp