(I just noticed this is not quite in a working state due to some linking problems, but I’ll add this here anyway.)

This replaces the local alignment with SSW with a hybrid approach where
- WFA2-lib is used to globally align the NAM
- ksw2 is used twice to extend the NAM towards the 5' and 3' end of the read

There are also a lot of restructuring commits in here to make the way the various alignment functions are called more consistent with each other.

This needs some optimizations. In particular, I temporarily removed the optimization of trying a gapfree alignment first (with `hamming_align`). However, the speed is roughly the same as before, in particular when setting match scores to 0 (which makes WFA2 faster).

## To Do

- [x] Fix the linking problem (`libwfa2.so` not found – switch from dynamic to static linking)
- [x] Put back the `hamming_align` optimization
- [ ] Run benchmarks
- [ ] Fix the macOS build error
- [ ] Replace SSW in rescue alignment with ksw2