@marcelm 2023-02-17
Finding long indels in the extensions may be a matter of tuning the scoring function. ksw2 supports a dual gap cost function that is supposed to help retain long gaps, maybe we could switch that on. Actually, now that I think about it, if we use the same scoring function as before, then we should actually be able to span the same indels as before. But there’re probably cases I haven’t thought about.

I just did some preliminary benchmarks for this PR and noticed that accuracy is slightly lower than before (~0.02 percentage points). This may be due to soft clipping, which happens more often now. Many alignments change from something like this: `4=1X95=` to this: `5S95=`. Using the current scores, any mismatch 4 bases away from the end of the read is going to be soft clipped. BWA-MEM has a penalty for soft clipping to avoid this, and ksw2 appears to have a way to give a bonus to alignments that reach the end of the query, which would have the same effect. This is related to #54.