@marcelm 2023-02-21
I added back the `hamming_align` optimization and with this, mapping appears to be a little bit faster than before. I still need to run more thorough benchmarks.