@marcelm 2023-02-23
Here’s a summary of my findings:
- Accuracy drops by 0.01 percentage points across all datasets (it was 0.02 before I added back the `hamming_align` optimization)
- The speedup ranges from no speedup at all for short reads to 1.5 times for longer reads

## Accuracy

dataset | main | this PR | accuracy reduction |
-|-|-|-|
drosophila/2x50-1M | 90.0643 | 90.0591 | 0.0052
drosophila/2x100-1M | 92.38105 | 92.37705 | 0.004
drosophila/2x150-1M | 93.2202 | 93.21225 | 0.00795
drosophila/2x200-1M | 93.506 | 93.4974 | 0.0086
drosophila/2x300-1M | 95.3651 | 95.35735 | 0.00775
CHM13/2x50-1M | 90.46475 | 90.4603 | 0.00445
CHM13/2x100-1M | 93.22665 | 93.2179 | 0.00875
CHM13/2x150-1M | 94.1315 | 94.11335 | 0.01815
CHM13/2x200-1M | 94.41465 | 94.4028 | 0.01185
CHM13/2x300-1M | 95.62665 | 95.616 | 0.01065
maize/2x50-1M | 71.1893 | 71.1803 | 0.009
maize/2x100-1M | 87.1095 | 87.10545 | 0.00405
maize/2x200-1M | 92.91535 | 92.9083 | 0.00705
maize/2x300-1M | 96.71805 | 96.7116 | 0.00645

## Speed

This was run on Drosophila only, reporting minimum runtime of three runs.

dataset | runtime before (s) | runtime after (s) | speedup
-|-|-|-
2x50bp | 49.660 | 49.298 | 1.01 |
2x100 bp | 64.250 | 60.817 | 1.06 |
2x150 bp | 78.750 | 67.922 | 1.16 |
2x200 bp | 100.810 | 78.441 | 1.29 |
2x300 bp | 199.235 | 131.731 | 1.51 |

Other notes:
- Alignment is not a bottleneck for short reads. The percentage of total runtime spent on it in main was ca. 19% (2x100 bp reads), so this limits how much total speedup we can have by making alignment faster. (This PR reduces the percentage to 14%).
- For 2x200 bp reads, time spent on alignment drops from ~32% to ~9%

I’d like to look into what happens to some of the reads that now get aligned incorrectly, maybe we can avoid the drop in accuracy.