@ksahlin 2023-02-23
To me this looks like great news! Although I would have to rerun the SNP/indel calling pipeline to know how/if it deteriorates around indels. Two notes (pulling in different directions):

1. I think extension time takes up a larger fraction of the time on smaller genomes, so the relative speedup will not be as high on larger genomes (for simulated data).
2. However, alignment is invoked less often in simulated reads compared to biological datasets (for some reason) as seen in [figure 2C](https://genomebiology.biomedcentral.com/articles/10.1186/s13059-022-02831-7/figures/2), so I suspect the speedup will be more prevalent in biological data.

Also, it could be worth a short discussion (at office) on how this partitioned alignment work. Perhaps after I get result from the indel calling benchmark.

Would you like me to start the SNP/indel benchmarks for this commit? (or tell me otherwise when you believe you have something ready to test).